data validation
The sign up now checks to see if the username and/or email exists in the DB currently. (if it does it will prompt the user with an error message. if it doesn't it will then insert the user into the DB). The sign up now also verifies that the passwords length is greater than 6 characters, it also checks if it contains and uppercase letter a lowercase letter and a symbol. Fixed bugs with log in where if the user miss spelt their email it would crash. I also created methods to create pop-up modals and send them to the render template.