sessions
The app now checks if a session has been created before continuing, if the user isn't logged in they are redirected to log in page. sessions now work and function on client side.
Session cookies for client side are not working yet so you'll have to log in each time.